+++ client/src/components/header/WikiInputField.tsx
@@ -1,26 +1,46 @@
-import React, { FormEvent } from 'react';
+import React, { useEffect } from 'react';
 import { ReactComponent as SearchCircle } from '@assets/image/search-circle-secondary.svg';
 import useInput from '@hooks/useInput';
 import { useNavigate } from 'react-router-dom';
 import URLS from '@constants/urls';
 import { twMerge } from 'tailwind-merge';
+import useGetDocumentSearch from '@api/get/useGetDocumentSearch';
+import useDebounce from '@hooks/useDebounce';
+import RelativeSearchTerms from './RelativeSearchTerms';
 
 interface WikiInputProps {
   id?: string;
   className?: string;
-  handleSubmit: (e: FormEvent) => void;
+  handleSubmit: (e: React.FormEvent) => void;
 }
 
 const WikiInputField = ({ id, className, handleSubmit }: WikiInputProps) => {
   const [value, setValue] = useInput<string>('');
   const navigate = useNavigate();
 
-  const onSubmit = (event: FormEvent) => {
+  const { titles, searchDocuments } = useGetDocumentSearch(value);
+  const searchDocumentsIfValid = () => {
+    if (value.trim() !== '' && /^[가-힣()0-9]*$/.test(value)) searchDocuments();
+  };
+
+  const debouncedSearchDocuments = useDebounce(searchDocumentsIfValid, 200);
+
+  useEffect(() => {
+    debouncedSearchDocuments();
+  }, [value]);
+
+  const onSubmit = (event: React.FormEvent, search?: string) => {
     event.preventDefault();
-    // eslint-disable-next-line no-unused-expressions
     if (value.trim() === '') return;
 
-    navigate(`${URLS.WIKI}/${value}`);
+    if (search !== undefined) {
+      navigate(`${URLS.WIKI}/${search}`);
+    } else if (titles !== undefined && titles.length !== 0) {
+      navigate(`${URLS.WIKI}/${titles[0]}`);
+    } else {
+      navigate(`${URLS.WIKI}/${value}`);
+    }
+
     setValue('');
     handleSubmit(event);
   };
@@ -29,7 +49,7 @@ const WikiInputField = ({ id, className, handleSubmit }: WikiInputProps) => {
     <form
       onSubmit={onSubmit}
       className={twMerge(
-        'flex h-11 px-4 py-2.5 rounded-xl bg-white border-grayscale-200 border-solid border gap-2',
+        'flex relative h-11 px-4 py-2.5 rounded-xl bg-white border-grayscale-200 border-solid border gap-2',
         className,
       )}
     >
@@ -43,6 +63,7 @@ const WikiInputField = ({ id, className, handleSubmit }: WikiInputProps) => {
       <button>
         <SearchCircle className="cursor-pointer max-[768px]:hidden" />
       </button>
+      {value.trim() !== '' && <RelativeSearchTerms searchTerms={titles ?? []} onSubmit={onSubmit} />}
     </form>
   );
 };
